Add files via upload
Update 4 of project.  Among other things, I have changed select and makeOrder in order to make files along with being able to get input.  I've separated the make order from select items as upon looking back, I saw that we did not need to create an order after selecting and instead just stored it in customer.  I've also separated the Bank check into verifying the overall number as upon asking the TA, he said we needed to verify it.  Finally, along with adding a file, I made it to where a supplier would also need to be listed as a supplier can only access a specific delivery order, not any one.